Added Users and course fields to training
I added two fields to the training course and users. Now all I have to
do is a populate and I can bring that data in no issues. I updated my
seeds folder to add a user to the users field and to also push courses
to the course field.