training delete and update rows
I fixed the training model I removed the course and user fields because
it was a lot of redundant data. I updated the seeds file to not add to
the course and users fields. I modified the home view to be cleaner and
make more sense. I updated the editTraining view with a delete button. I
edit the show view.